refactor: Move supporting directories to src directory
Move simulator, examples, and docs directories from root to src/ to continue project restructuring. Fixed import paths in example files to align with new directory structure. Confirmed tokyo-py virtual environment is properly set up and excluded from reorganization.
