+++ src/docs/ai/README.md
