+++ src/docs/ai/companion-ai-design.md
