+++ src/docs/ai/personality-engine.md
