+++ src/docs/context_data/agent-prompt-vector-store.md
