+++ src/docs/design/game-design-doc.md
