+++ src/docs/design/scenario-detection.md
