+++ src/docs/design/system/bedrock_integration_implementation.md
