+++ src/docs/design/system/ollama_integration_implementation.md
