+++ src/docs/design/system/personality_engine_implementation.md
