+++ src/docs/design/system/scenario_detection_implementation.md
