+++ src/docs/design/system/template_system_implementation.md
