+++ src/docs/pm/companion-ai-implementation-plan.md
